feat: update cdk to v2
This also updates the npm lockfile to v2 and uses Node 16 to build
since that ships with NPM that supports lockfile v2.

Handle changes to Cloud Manifest currently used.

This is not a breaking change since we don't have a visible dependency
on CDK in the produced library. It will support Cloud Assemblies from
both v1 and v2 for now.